SOL-151468: move retryState isolation test to test/integration
test/integration/README.md defines the tier boundary: unit tests live
next to their component and exercise one component's public surface,
integration tests live in test/integration/ and exercise "at least two
internal/ components composed together" for invariants that only make
sense once wired up — the README calls out isolation and concurrency by
name.

The previous placement in internal/semp/resilience/sender_test.go
followed local convention (other 401 tests live there) but violated the
tier rule. The invariant being asserted — per-request retryState
isolation across concurrent Do() calls — is meaningless without a real
auth.BasicAuthenticator wired to a shared resilience.SafeCookieJar: the
Sender declares the retry-once cap, BasicAuthenticator declares the
recovery strategy, and only under concurrent composition does the
observable failure ("only one goroutine's retry fires") emerge. Two
internal/ components composed under load — bullseye for the integration
tier.

The moved file uses only exported APIs (resilience.New,
resilience.NewSafeCookieJar, resilience.NewSemaphore,
auth.NewBasicAuthenticator) — no test-only helpers cross the package
boundary. The Sender is constructed directly rather than routed through
BrokerPool: the invariant lives inside resilience.checkRetry, and going
through the full pool/SEMPv1 stack would add parsing, retry-safe
marking, and correlation-header layers that are irrelevant to the
invariant and would cloud diagnosis on failure.

Filename follows the README convention (name after the invariant, not
the components): retry_state_isolation_test.go, matching the sibling
broker_credential_isolation_test.go.
